Enable touchpad history navigation in linux as default
Prior to this commit, touchpad history navigation was disabled by
default in Linux. The feature could be enabled with the command line
argument --enable-features=TouchpadOverscrollHistoryNavigation. However,
it did not work as well as it did in Windows and ChromeOS, because the
navigation action was only performed when the touchpad was touched
again, not when the fingers were lifted from the touchpad because an
ET_SCROLL_GESTURE_END was missing. However, this problem has been fixed
with c4a12794a6a6ac17357528edb34cd31596bc880e so that this feature can
be enabled by default.

BUG=1315996

Change-Id: I34491398c00fc93dd4cd5356e1101c90a1e36c34